Local building is again.
Appears the C++ Game mode base got corrupted and the editor would no longer load the default maps.
I rebuilt new maps and character, and this build is able to be iterated with again.
Make sure to close and re-open the build before checking in just to make sure another editor blocking issue is not introduced.
Clock is not working correctly.