Added Cross-Platform Voice Chat Pro Plugin
Create a Plugins folder in the project's root folder and put the source code for the plugin there. Then right-clicked the GIGWorlds.uproject file and selected Generate Visual Studio project files option, open the solution in Visual Studio, and build the project on Visual Studio, then Opened the Unreal Engine Project and it built out the binaries.  Seems like everything built, now need to see if/how it works.
Reference:
https://answers.unrealengine.com/questions/518976/how-to-install-plugins-in-sourcebuild.html