#13 Invert the Mouse
Incorporated the Complete Menu System (https://www.unrealengine.com/marketplace/en-US/product/complete-menu-system)  into the project.  Documentation check in with source code.  Important to note, running a in editor debug build in the selected viewport with multiple clients created, causes a duplicated controller and menu error.  This can be avoided by running in standalone mode.  Note this is an in-progress checkin and further customization is ongoing.